requirements.txt
wheel==0.36.2
yoyo-migrations==7.2.1
psycopg2==2.9
tabulate==0.8.9 